requirements.txt
fastapi==0.95.2
pymongo==4.7.1
pydantic==1.10.13
python-dotenv==1.0.1
uvicorn[standard]==0.29.0
pytest==8.2.1
mongomock==4.1.2
httpx==0.27.0
setuptools==69.2.0 